refactor: удаление ~1800 строк мёртвого кода и перенос diarization.py
- Удалены неиспользуемые файлы: verify_llm_data_transfer.py, verify_od_template.py,
  simple_messages.py, validation_utils.py, smart_selection_improvements.md
- Удалена мёртвая функция generate_protocol_consolidated_simplified из llm_providers.py
- Удалены 8 мёртвых методов из database.py (pending messages, performance metrics)
- Удалены неиспользуемые BatchProcessor, optimize_io_operations, async_lru_cache
  из async_optimization.py
- Перенесён diarization.py → src/services/diarization_service.py
- Обновлены все связанные импорты и __init__.py
