+++ src/performance/async_optimization.py
@@ -229,54 +229,6 @@ class ThreadPoolManager:
         self.process_pool.shutdown(wait=True)
 
 
-class BatchProcessor:
-    """Пакетная обработка для оптимизации производительности"""
-    
-    def __init__(self, batch_size: int = 10, flush_interval: float = 5.0):
-        self.batch_size = batch_size
-        self.flush_interval = flush_interval
-        self.pending_items: List[Any] = []
-        self.last_flush = time.time()
-        self.processor_func: Optional[Callable] = None
-    
-    def set_processor(self, func: Callable):
-        """Установить функцию обработки пакета"""
-        self.processor_func = func
-    
-    async def add_item(self, item: Any) -> bool:
-        """Добавить элемент в пакет"""
-        self.pending_items.append(item)
-        
-        # Проверяем условия для флуша
-        should_flush = (
-            len(self.pending_items) >= self.batch_size or
-            time.time() - self.last_flush >= self.flush_interval
-        )
-        
-        if should_flush:
-            await self.flush()
-            return True
-        
-        return False
-    
-    async def flush(self):
-        """Обработать накопленный пакет"""
-        if not self.pending_items or not self.processor_func:
-            return
-        
-        items_to_process = self.pending_items.copy()
-        self.pending_items.clear()
-        self.last_flush = time.time()
-        
-        try:
-            await self.processor_func(items_to_process)
-            logger.debug(f"Обработан пакет из {len(items_to_process)} элементов")
-        except Exception as e:
-            logger.error(f"Ошибка обработки пакета: {e}")
-            # Возвращаем элементы обратно для повторной попытки
-            self.pending_items.extend(items_to_process)
-
-
 class AsyncResourceManager:
     """Менеджер асинхронных ресурсов"""
     
@@ -363,66 +315,3 @@ async def optimized_file_processing():
 # Глобальные экземпляры
 task_pool = AsyncTaskPool(max_concurrent=5)  # Консервативное значение для начала
 thread_manager = ThreadPoolManager()
-
-
-async def optimize_io_operations(operations: List[Callable]) -> List[Any]:
-    """Оптимизированное выполнение I/O операций"""
-    if not operations:
-        return []
-    
-    # Разделяем на пакеты для контроля нагрузки
-    batch_size = min(10, len(operations))
-    results = []
-    
-    for i in range(0, len(operations), batch_size):
-        batch = operations[i:i + batch_size]
-        batch_tasks = [op() for op in batch]
-        batch_results = await asyncio.gather(*batch_tasks, return_exceptions=True)
-        results.extend(batch_results)
-        
-        # Небольшая пауза между пакетами для снижения нагрузки
-        if i + batch_size < len(operations):
-            await asyncio.sleep(0.1)
-    
-    return results
-
-
-def async_lru_cache(maxsize: int = 128, ttl: float = 3600.0):
-    """Асинхронный LRU кэш с TTL"""
-    cache = {}
-    access_times = {}
-    
-    def decorator(func):
-        async def wrapper(*args, **kwargs):
-            # Создаем ключ кэша
-            key = f"{func.__name__}:{hash((args, tuple(sorted(kwargs.items()))))}"
-            
-            # Проверяем кэш
-            if key in cache:
-                # Проверяем TTL
-                if time.time() - access_times[key] < ttl:
-                    access_times[key] = time.time()
-                    return cache[key]
-                else:
-                    # Просрочен
-                    del cache[key]
-                    del access_times[key]
-            
-            # Выполняем функцию
-            result = await func(*args, **kwargs)
-            
-            # Сохраняем в кэш
-            if len(cache) >= maxsize:
-                # Удаляем самый старый элемент
-                oldest_key = min(access_times.keys(), key=access_times.get)
-                del cache[oldest_key]
-                del access_times[oldest_key]
-            
-            cache[key] = result
-            access_times[key] = time.time()
-            
-            return result
-        
-        return wrapper
-    
-    return decorator
