+++ src/utils/__init__.py
@@ -3,11 +3,9 @@
 """
 
 from .logging_utils import setup_logging
-from .validation_utils import validate_file_type, validate_file_size
 from .message_utils import send_long_message, format_error_message
 
 __all__ = [
     "setup_logging",
-    "validate_file_type", "validate_file_size", 
     "send_long_message", "format_error_message"
 ]
