+++ src/ux/__init__.py
@@ -4,7 +4,6 @@
 
 from .progress_tracker import ProgressTracker, ProgressFactory
 from .message_builder import MessageBuilder
-from .simple_messages import SimpleMessages
 from .feedback_system import (
     FeedbackCollector, FeedbackUI, QuickFeedbackManager, 
     feedback_collector, setup_feedback_handlers
@@ -15,7 +14,6 @@ __all__ = [
     "ProgressTracker",
     "ProgressFactory", 
     "MessageBuilder",
-    "SimpleMessages",
     "FeedbackCollector",
     "FeedbackUI",
     "QuickFeedbackManager",
